IntroProject/backend/requirements.txt
asgiref==3.4.1
Django==3.2.13
django-cors-headers==3.10.0
django-six==1.0.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
PyJWT==2.4.0
python-dotenv==0.19.1
pytz==2021.3
pyotp==2.6.0
sqlparse==0.4.2
whitenoise==5.3.0
xlwt==1.3.0
scipy==1.8.1

#for postgres
psycopg2-binary==2.8.6